# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Skorvia API - Brand credibility and ad-readiness analyzer
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List all analyses for the session/user
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Create a new brand analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisBody"
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a specific analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyses/{id}/run:
    post:
      operationId: runAnalysis
      tags: [analyses]
      summary: Trigger the AI analysis process for a brand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis started or completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CreateAnalysisBody:
      type: object
      properties:
        brandName:
          type: string
        websiteUrl:
          type: string
        instagramHandle:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        competitor1:
          type: ["string", "null"]
        competitor2:
          type: ["string", "null"]
        competitor3:
          type: ["string", "null"]
        industry:
          type: string
        email:
          type: ["string", "null"]
      required:
        - brandName
        - websiteUrl
        - industry
    Analysis:
      type: object
      properties:
        id:
          type: integer
        brandName:
          type: string
        websiteUrl:
          type: string
        instagramHandle:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        competitor1:
          type: ["string", "null"]
        competitor2:
          type: ["string", "null"]
        competitor3:
          type: ["string", "null"]
        industry:
          type: string
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, analyzing, completed, failed]
        overallScore:
          type: ["number", "null"]
        websiteScore:
          type: ["number", "null"]
        socialScore:
          type: ["number", "null"]
        contentScore:
          type: ["number", "null"]
        reviewsScore:
          type: ["number", "null"]
        competitorScore:
          type: ["number", "null"]
        adReadinessLevel:
          type: ["string", "null"]
          enum: [not_ready, getting_there, almost_ready, ready, null]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - brandName
        - websiteUrl
        - industry
        - status
        - createdAt
        - updatedAt
    AnalysisResult:
      type: object
      properties:
        analysis:
          $ref: "#/components/schemas/Analysis"
        insights:
          type: array
          items:
            $ref: "#/components/schemas/Insight"
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/ActionTask"
      required:
        - analysis
        - insights
        - tasks
    Insight:
      type: object
      properties:
        id:
          type: integer
        analysisId:
          type: integer
        category:
          type: string
          enum: [website, social_media, content, reviews, competitors, seo]
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, warning, info, positive]
        createdAt:
          type: string
      required:
        - id
        - analysisId
        - category
        - title
        - description
        - severity
        - createdAt
    ActionTask:
      type: object
      properties:
        id:
          type: integer
        analysisId:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: integer
        category:
          type: string
          enum: [website, social_media, content, reviews, seo, competitors, branding]
        estimatedDays:
          type: integer
        isCompleted:
          type: boolean
        completedAt:
          type: ["string", "null"]
        isDailyTask:
          type: boolean
        scheduledDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - analysisId
        - title
        - description
        - priority
        - category
        - estimatedDays
        - isCompleted
        - isDailyTask
        - createdAt
...